fling-core/pyproject.toml
[tool.poetry]
name = "fling-core"
version = "0.1.2"
description = "Fling core components for Side Project Management"
authors = ["Joshua McKenty <jmckenty@gmail.com>", "Anouk Ruhaak <anoukruhaak@gmail.com>"]
readme = "README.md"
packages = [{include = "fling_core"}]

[tool.poetry.dependencies]
python-dotenv = "^1.0.0"
dynaconf = "^3.1.12"
python = "^3.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"